apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[project]
name = "studynest-api"
version = "0.1.0"
description = "StudyNest cloud API + local AI sidecar"
# Code uses PEP 604 unions (`str | None`) and PEP 585 generics (`list[str]`) → 3.10+.
# Local AI wheels / deps are easiest on 3.11–3.12 — use those if possible.
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "httpx>=0.27",
    "supabase>=2.7.0",
    "python-multipart>=0.0.9",
    "ulid-py>=1.1.0",
]

[project.optional-dependencies]
local-ai = [
    "llama-cpp-python>=0.3.1",
]
cloud-ai = [
    "google-generativeai>=0.7.2",
    "openai>=1.40.0",
]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
